+++ enum_ext_gen/lib/src/generators/enum_ext_generator.dart
@@ -40,6 +40,8 @@ class EnumExtGenerator extends GeneratorForAnnotation<EnumExt> {
 
     buffer.writeln("}");
 
+    buffer.write(generateEnumByCustomPropertySelector(enumName, fields));
+
     return buffer.toString();
   }
 }